requirements.txt
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
lxml==5.3.1
MarkupSafe==3.0.2
nltk==3.9.1
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pycparser==2.22
pypdfium2==4.30.1
python-docx==1.1.2
regex==2024.11.6
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
